refactor: various changes
* Split leaf, subtree and root into multiple files
  + spec: all the spec code
  + running_<node>: all the code used for running processes (e.g. termination)
  + opts: all options used on the public API

* Add tests for one for one permanent restarts

* Rename TerminationError to TerminationMessage
