+++ src/tests/restart.rs
@@ -1,0 +1,1 @@
+mod permanent_tests;
